Add Book and Course data management features
- Created `BookData` class to manage book data from `bookdata.json`.
- Updated `CourseData` to use `coursedata.json` for course information.
- Added `bookdata.json` and `coursedata.json` files with structured data.
- Implemented endpoints in `BookEndPoints` and `CourseEndPoints` for data retrieval.
- Renamed `Book` class to `BookModel` and updated properties.
- Changed `courseLengthInHours` type to `double` for precision.
- Registered `BookData` service in the dependency injection container.

+++ fs-2025-a-api-demo-002/Models/Course.cs
@@ -8,7 +8,7 @@
         public string courseType { get; set; }
         public string courseName { get; set; }
         public int courseLessonCount { get; set; }
-        public int courseLengthInHours { get; set; }
+        public double courseLengthInHours { get; set; }
         public string shortDescription { get; set; }
         public string courseImage { get; set; }
         public int priceInUSD { get; set; }
